Reestructurado el código:
-añadido la carpeta clients.
- Renombrada las clases y modulos y metodos de ollama utils(ollama_client) y llama utilis (llama_models)
-movidas a cleints
