+++ monitorsystem.py
@@ -2,9 +2,8 @@ from threading import Thread, Event
 from datetime import datetime
 from metrics.hardware_metrics import HardwareMetrics as hm
 from metrics.promptmetrics import PromptMetrics as pm
-from utils.ollama_utils import ollama_model_checker,ollamaClient
-from utils.llama_utils import LlamaModels
-from utils.configuration import client_default_ollama as cdo 
+from clients.ollama_client import ollama_model_checker,Ollamaclient
+from clients.llama_model import LlamaModels
 from utils.configuration import TIME_BETWEEN_MODELS as tbm, TIME_BETWEEN_PROMPTS as tbp, OLLAMA_MODEL_LIST as oml, PROMPT_LIST as pl
 import time
 
@@ -23,7 +22,7 @@ def main_ollama():
         hardware_metric_filepath =f"results/ollama_hardware_metrics_ollama_{current_time}_{model}.csv"
         pm.create_csv_file(prompt_metric_filepath)
         hm.create_csv_file(hardware_metric_filepath)
-        oc = ollamaClient()
+        oc = Ollamaclient()
         
         for i in range (len(pl)):
            
@@ -35,7 +34,7 @@ def main_ollama():
 
             event = Event()
             prompt_thread = Thread(
-                target=oc.ollama_query_and_save_with_event,
+                target=oc.query_event_save,
                 args=(pl[i], model,prompt_metric_filepath,event,i)
             )
             hardware_thread = Thread(
